Consolidate functions
Moved function from sort_utils.py to resize_images.py since it's only used
there, removed sort_utils.py
